requirements.txt
Flask==2.1.1
pymongo==4.0.1
face_recognition==1.3.0
PyJWT==2.1.0
Werkzeug==2.1.1
